Find position by calculating roots of polynomial
For cases where alpha and beta are linearized, delta_rho becomes a
relatively low-degree polynomial (order two or three. The rpoly
algorithm described by Jenkins and Traub efficiently and accurately
calculates all roots of the polynomial. The root closest to the bottom
(closest to 1.) is chosen as the position of the neutral surface.